pyproject.toml
[project]
name = "reachy-mini-agent"
version = "0.1.0"
description = "Reachy Mini agent - Clawdbot brain with ElevenLabs voice and Honcho memory"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    # Robot SDK
    "reachy-mini>=1.2.3",
    # AI services
    "httpx>=0.27.0",
    "openai>=1.0.0", # For Whisper STT
    "pydub>=0.25.0", # Audio format conversion
    # Vision
    "mediapipe>=0.10.0",
    "opencv-python>=4.8.0",
    # "face-recognition>=1.3.0",  # Optional - requires cmake/dlib
    # Memory
    "honcho-ai>=2.0.0",
    # Core
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "requests>=2.31.0",
    "face-recognition-models",
    "gst-signalling>=1.1.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["reachy_agent"]

[tool.uv]
dev-dependencies = []

[tool.uv.sources]
face-recognition-models = { git = "https://github.com/ageitgey/face_recognition_models" }
